Gorgon.Graphics.Core
- Finalized the resource copier functionality, and documented it.
- Added internal state reset for command lists when pulling from the pool.
- Fixed bug in megabuffer when alignment is 0.
- Added functionality to the barrier manager to ensure state compatibility between queues.

+++ Gorgon/Gorgon.Graphics/Core/GorgonGraphics.cs
@@ -914,7 +914,7 @@ public unsafe sealed class GorgonGraphics
     /// </remarks>
     internal void InitializeTexture(GorgonTexture texture)
     {
-        UploadHeaps.Allocate((ulong)texture.SizeInBytes, (int)texture.Info.Alignment, out CpuBufferAllocation allocation);
+        UploadHeaps.Allocate((ulong)texture.SizeInBytes, texture.Info.Alignment, out CpuBufferAllocation allocation);
 
         Debug.Assert(allocation.IsAvailable, $"Could not allocate upload memory for texture '{texture.Name}'");
